split EditMap and VisuMap files
