+++ components/map/EditMap.tsx
@@ -5,14 +5,14 @@ import 'maplibre-gl/dist/maplibre-gl.css';
 import { useMap } from '@/components/map/useMap';
 import { useMapLayers } from '@/components/map/useMapLayers';
 import { useMapControls } from '@/components/map/useMapControls';
-import { useVisuMapEvents } from '@/components/map/useVisuMapEvents';
+import { useEditionMapEvents } from '@/components/map/useEditionMapEvents';
 import { useMapStateSync } from '@/components/map/useMapStateSync';
-import { useMapStateSyncSelectedBuilding } from '@/components/map/useMapStateSyncSelectedBuilding';
 import {
   MapLayer,
   MapBackgroundLayer,
   MapBuildingsLayer,
 } from '@/stores/map/map-slice';
+import { useMapEditBuildingShape } from '@/components/map/useMapEditBuildingShape';
 
 type Props = {
   disabledLayers?: MapLayer[];
@@ -20,22 +20,23 @@ type Props = {
   defaultBuildingLayer?: MapBuildingsLayer;
 };
 
-export default function VisuMap({
+export default function EditMap({
   disabledLayers,
   defaultBackgroundLayer,
   defaultBuildingLayer,
 }: Props) {
   const { map, mapContainer } = useMap({ disabledLayers });
+
   useMapLayers({
     map,
     defaultBackgroundLayer,
     defaultBuildingLayer,
-    selectedBuildingisGreen: true,
+    selectedBuildingisGreen: false,
   });
-  useMapControls(map);
-  useVisuMapEvents(map);
-  useMapStateSync(map);
-  useMapStateSyncSelectedBuilding(map);
 
+  useMapControls(map);
+  useEditionMapEvents(map);
+  useMapStateSync(map);
+  useMapEditBuildingShape(map);
   return mapContainer;
 }
